Bug 1913666 - Part 1: Introduce Fullscreen Service r=edgar!
Introduces the fullscreen service which is just a skeleton implementation in this patch. In later patches
the fullscreen service will manage a collection of "fullscreen managers" which control
fullscreen requests on a per-app-window basis, since Firefox allows multiple windows
to go DOM fullscreen, but only 1 tab may be in DOM Fullscreen per window.

Differential Revision: https://phabricator.services.mozilla.com/D301185