Bug 1999663 - Fix restoration of scroll position r=#dom-core!
This patch updates the scroll position on new loads (via
nsDocShell::InternalLoad), which makes sure that the scroll position
data is up to date.

It also addresses the fact that we can't (at the moment) use
mActiveEntry in nsDocShell to restore scroll position, because when
restoring of scroll happens, the active session history entry has not
been updated yet.

I'm unclear if this is a spec issue or if we are doing something wrong,
but going forward, but when such a fix may or may not arrive, aligning
this fix with that, will be trivial, as it means just passing the
mActiveEntry as the "target entry" instead (this should
continue to work, anyway though).

Differential Revision: https://phabricator.services.mozilla.com/D272446